Introduced the CharacterMap class to the Game class structure.
This class organizes characters much like the Map class organizes MapTiles, keeping track of character's coordinates and IDs. The Player is basically now abstracted as simply the character with ID = 0. The Player class no longer contains its own coordinates.

The purpose of these changes is to pave the way for NPCs ("monsters," if you will). Right now, the only Character is the Player, but eventually a real Character class will be added and both the Player and NPC class will extend it.