InputManager replaces InputProcessor.
This new module establishes the basis of the "advanced input system" (issue #5). The KeymapInputManager manages the Keymap classes. Keymaps dont define keys:commands like in the old system; they now define key:actions mappings. Actions are anonymous functions that take two params: a KeymapInputManager and a Keymap. When the KeymapInputManager is running, it will run the actions for the given key in order, supplying itself and the current keymap as arguments. This allows for actions to do things like switch between keymaps, return results from keymaps, and buffer partial-commands.

This system is much more robust/flexible than the old way. The throw command has been fully implemented in the InputManager but isn't finished in the Engine/DisplayManager yet.

~ Out with the old and in with the new~